+++ src/helpers/functions/Card.js
@@ -1,0 +1,5 @@
+import cardCategories from "../constants/cardCategories.json";
+
+export function getCardCategories() {
+    return JSON.parse(JSON.stringify(cardCategories));
+}